# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError:
    properties:
      error:
        type: string
    type: object
...
  internal_chat.CreateRoomRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
      visibility:
        allOf:
        - $ref: '#/definitions/internal_chat.RoomVisibility'
        enum:
        - public
        - private
    required:
    - name
    type: object
...
  internal_chat.JoinResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_chat.Message:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      room_id:
        type: integer
      user_id:
        type: integer
      username:
        description: joined from users table
        type: string
    type: object
  internal_chat.MessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/internal_chat.Message'
        type: array
    type: object
  internal_chat.Room:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/internal_chat.RoomRole'
      visibility:
        $ref: '#/definitions/internal_chat.RoomVisibility'
    type: object
...
  internal_chat.RoomRole:
    enum:
    - owner
    - moderator
    - member
    type: string
    x-enum-varnames:
    - RoomRoleOwner
    - RoomRoleModerator
    - RoomRoleMember
  internal_chat.RoomVisibility:
    enum:
    - public
    - private
    type: string
    x-enum-varnames:
    - RoomVisibilityPublic
    - RoomVisibilityPrivate
  internal_chat.RoomsResponse:
    properties:
      rooms:
        items:
          $ref: '#/definitions/internal_chat.Room'
        type: array
    type: object
  internal_chat.SendMessageRequest:
    properties:
      content:
        maxLength: 4000
        minLength: 1
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the Gogo DL real-time chat backend.
  title: Gogo DL API
  version: "1.0"
paths:
...
  /api/v1/rooms:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.RoomsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: List rooms visible to the current user
      tags:
      - rooms
    post:
      consumes:
      - application/json
      parameters:
      - description: Room payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_chat.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_chat.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a room
      tags:
      - rooms
  /api/v1/rooms/{id}:
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get a room visible to the current user
      tags:
      - rooms
...
  /api/v1/rooms/{id}/join:
    post:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.JoinResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Join a public room
      tags:
      - rooms
...
  /api/v1/rooms/{id}/messages:
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - default: 50
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Return messages with IDs below this cursor
        in: query
        minimum: 1
        name: before
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.MessagesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: List room messages
      tags:
      - messages
    post:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_chat.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_chat.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Send a message to a room
      tags:
      - messages
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the access token as: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"